requirements.txt
attrs==21.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
fake-useragent==0.1.11
greenlet==1.1.0
idna==2.10
iniconfig==1.1.1
numpy==1.21.0
packaging==21.0
pandas==1.2.5
pluggy==0.13.1
psycopg2==2.9.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.20
toml==0.10.2
urllib3==1.26.6
